+++ src/main/java/com/bob/mate/domain/chat/controller/ChatController.java
@@ -1,16 +1,7 @@
 package com.bob.mate.domain.chat.controller;
 
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
 
-@Controller
-@Slf4j
+@RestController
 public class ChatController {
-
-    @GetMapping("/chat")
-    public String getChat(){
-        log.info("@ChatController.GET()");
-        return "chat";
-    }
 }
