+++ src/main/java/com/bob/mate/domain/post/repository/PostRepository.java
@@ -1,6 +1,6 @@
 package com.bob.mate.domain.post.repository;
 
-import com.bob.mate.domain.post.domain.Post;
+import com.bob.mate.domain.post.entity.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostRepository extends JpaRepository<Post, Long>, PostRepositoryCustom{
