+++ src/main/java/com/bob/mate/domain/post/repository/PostRepositoryCustom.java
@@ -5,6 +5,5 @@ import com.bob.mate.domain.post.dto.PostQueryDto;
 import java.util.List;
 
 public interface PostRepositoryCustom {
-
     List<PostQueryDto> findAllPost();
 }
