+++ src/main/java/com/bob/mate/domain/post/repository/PostRepositoryImpl.java
@@ -1,6 +1,6 @@
 package com.bob.mate.domain.post.repository;
 
-import com.bob.mate.domain.post.domain.QTag;
+import com.bob.mate.domain.post.entity.QTag;
 import com.bob.mate.domain.post.dto.PostQueryDto;
 import com.bob.mate.domain.post.dto.QPostQueryDto;
 import com.querydsl.jpa.impl.JPAQueryFactory;
@@ -8,8 +8,8 @@ import lombok.RequiredArgsConstructor;
 
 import java.util.List;
 
-import static com.bob.mate.domain.post.domain.QPost.post;
-import static com.bob.mate.domain.post.domain.QTag.*;
+import static com.bob.mate.domain.post.entity.QPost.post;
+import static com.bob.mate.domain.post.entity.QTag.*;
 
 @RequiredArgsConstructor
 public class PostRepositoryImpl implements PostRepositoryCustom{
