+++ src/main/java/com/bob/mate/domain/user/entity/User.java
@@ -1,6 +1,6 @@
 package com.bob.mate.domain.user.entity;
 
-import com.bob.mate.domain.post.domain.Post;
+import com.bob.mate.domain.post.entity.Post;
 import com.bob.mate.global.audit.AuditListener;
 import com.bob.mate.global.audit.Auditable;
 import com.bob.mate.global.audit.TimeEntity;
